+++ package.json
@@ -67,6 +67,7 @@
     "babel-core": "7.0.0-bridge.0",
     "babel-eslint": "10.0.3",
     "babel-loader": "8.0.6",
+    "babel-plugin-styled-components": "1.12.0",
     "babel-plugin-transform-react-remove-prop-types": "0.4.24",
     "browser-sync": "2.26.7",
     "chalk": "2.4.2",
