+++ src/components/Legend/Legend.styles.js
@@ -5,9 +5,8 @@ export const LegendList = styled.div`
   display: flex;
   flex-direction: row;
   flex-wrap: wrap;
-  border: 1px #eaeaea solid;
   border-radius: 2px;
-  padding: 6px 12px;
+  padding: 6px 3px;
 `;
 
 export const LegendItem = styled.div`
